+++ t/20_cache_full.t
@@ -14,7 +14,16 @@ my $V = $ENV{INTEGRITY};        # Verbose integrity checking?
 print "1..111\n";
 
 my $N = 1;
-use Tie::File::FixedRecLen;
+BEGIN {
+    eval {require Tie::File::FixedRecLen};
+
+    if ($@) {
+      print "1..0 # skipped... cannot use Tie::File::FixedRecLen with your version of Tie::File
+";
+      exit;
+    }
+}
+
 print "ok $N\n"; $N++;
 
 open F, "> $file" or die $!;
@@ -225,7 +234,9 @@ sub ctrlfix {
 END {
   undef $o;
   untie @a;
-  1 while unlink $file;
+  if (defined $file) {
+      1 while unlink $file;
+  }
 }
