+++ t/31_autodefer.t
@@ -19,7 +19,16 @@ my ($o, $n, @a);
 print "1..65\n";
 
 my $N = 1;
-use Tie::File::FixedRecLen;
+BEGIN {
+    eval {require Tie::File::FixedRecLen};
+
+    if ($@) {
+      print "1..0 # skipped... cannot use Tie::File::FixedRecLen with your version of Tie::File
+";
+      exit;
+    }
+}
+
 print "ok $N\n"; $N++;
 
 open F, "> $file" or die $!;
@@ -180,6 +189,8 @@ sub ctrlfix {
 END {
   undef $o;
   untie @a;
-  1 while unlink $file;
+  if (defined $file) {
+      1 while unlink $file;
+  }
 }
