+++ t/32_defer_misc.t
@@ -17,7 +17,16 @@ my ($o, $n);
 print "1..53\n";
 
 my $N = 1;
-use Tie::File::FixedRecLen;
+BEGIN {
+    eval {require Tie::File::FixedRecLen};
+
+    if ($@) {
+      print "1..0 # skipped... cannot use Tie::File::FixedRecLen with your version of Tie::File
+";
+      exit;
+    }
+}
+
 print "ok $N\n"; $N++;
 
 open F, "> $file" or die $!;
@@ -230,6 +239,8 @@ sub ctrlfix {
 END {
   undef $o;
   untie @a;
-  1 while unlink $file;
+  if (defined $file) {
+      1 while unlink $file;
+  }
 }
